+++ sdk/monitor/azure-opentelemetry-exporter-azuremonitor/CHANGELOG.md
@@ -8,6 +8,10 @@
 - Renamed artifact to `azure-opentelemetry-exporter-azuremonitor`.
 - Replaced `instrumentationKey()` with `connectionString()` in the `AzureMonitorExporterBuilder`.
 
+### Dependency updates
+- Updated `opentelemetry-api` and `opentelemetry-sdk` version to `0.13.1` 
+  and `azure-core-tracing-opentelemetry` to `1.0.0-beta.7`
+
 ## 1.0.0-beta.1 (2020-10-06)
 
 ### New Features